172719 resolve review comments AB#172719
