+++ src/UKHO.S100PermitService.API/Controllers/PermitController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using UKHO.S100PermitService.Common.Enum;
+using UKHO.S100PermitService.Common.Utilities;
 
 namespace UKHO.S100PermitService.API.Controllers
 {
