+++ src/UKHO.S100PermitService.Common/Enum/EventIds.cs
@@ -1,7 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using Microsoft.Extensions.Logging;
-
-namespace UKHO.S100PermitService.Common.Enum
+﻿namespace UKHO.S100PermitService.Common.Enum
 {
     /// <summary>
     /// Event Ids
@@ -9,35 +6,18 @@ namespace UKHO.S100PermitService.Common.Enum
     public enum EventIds
     {
         /// <summary>
-        /// 940001 - An unhandled exception occurred while processing the request.
+        /// 840001 - An unhandled exception occurred while processing the request.
         /// </summary>
-        UnhandledException = 940001,
+        UnhandledException = 840001,
 
         /// <summary>
-        /// 940002 - Generate Permit API call started.
+        /// 840002 - Generate Permit API call started.
         /// </summary>
-        GeneratePermitStarted = 940002,
+        GeneratePermitStarted = 840002,
 
         /// <summary>
-        /// 940003 - Generate Permit API call end.
+        /// 840003 - Generate Permit API call end.
         /// </summary>
-        GeneratePermitEnd = 940003
-    }
-
-    /// <summary>
-    /// EventId Extensions
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    public static class EventIdExtensions
-    {
-        /// <summary>
-        /// Event Id
-        /// </summary>
-        /// <param name="eventId"></param>
-        /// <returns></returns>
-        public static EventId ToEventId(this EventIds eventId)
-        {
-            return new EventId((int)eventId, eventId.ToString());
-        }
-    }
+        GeneratePermitEnd = 840003
+    }   
 }
